+++ src/main/java/services/fetchs/fetch/Fetch.java
@@ -17,6 +17,7 @@ public class Fetch<T> implements IFetch<T> {
    Connection _connection = SqliteConnect.getConnect();
    Statement _statement = _connection.createStatement();
    ResultSet _resultSet = _statement.executeQuery(sqlSelect);
+
    PreData<T> _PreData = new PreData<>(_resultSet);
 
    return _PreData.setupResult(object);
