--- src/main/java/services/fetchs/ifetch/IFetchAll.java
@@ -1,8 +1,0 @@
-package services.fetchs.ifetch;
-
-import java.util.List;
-
-public interface IFetchAll<T> {
-
-    List<T> foundAll();
-}
