--- src/main/java/services/fetchs/ifetch/IFetchById.java
@@ -1,7 +1,0 @@
-package services.fetchs.ifetch;
-
-import java.util.List;
-
-public interface IFetchById<T> {
-    List<T> found(long id);
-}
