--- src/main/java/services/fetchs/ifetch/IFetchByQuery.java
@@ -1,8 +1,0 @@
-package services.fetchs.ifetch;
-
-import java.util.List;
-
-public interface IFetchByQuery<T> {
-    List<T> found(String query);
-
-}
